change architectur a lil bit
